backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "bert-score==0.3.13",
    "celery==5.4.0",
    "datasets==2.11",
    "django==5.0.6",
    "django-cors-headers==4.3.1",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt==5.3.1",
    "evaluate[evaluator]==0.4.2",
    "httpx==0.27.2",
    "nltk==3.8.1",
    "nmslib-metabrainz>=2.1.3",
    "numpy>=1.26.4",
    "openai==1.57.4",
    "pandas==2.2.2",
    "prettytable==3.10.0",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "rank-bm25>=0.2.2",
    "redis==5.0.7",
    "requests>=2.32.3",
    "rouge-score==0.1.2",
    "scispacy>=0.2.4",
    "sendgrid==6.11.0",
    "sentencepiece==0.2.0",
    "simplejson==3.19.2",
    "spacy>=3.7.5",
    "sqlparse==0.5.0",
    "together==1.3.2",
    "torch==2.3.1",
    "tqdm>=4.67.1",
    "transformers==4.36.2",
]
